test_deployment_v2/requirements.txt
# LP Strategy v2 Dependencies - Complete list for fly.io deployment

# Solana blockchain
solana>=0.34.0
solders>=0.21.0
anchorpy>=0.19.0

# HTTP client
httpx>=0.25.0
aiohttp>=3.9.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Base58 encoding
base58>=2.1.0

# Logging
structlog>=23.2.0

# Retry logic
tenacity>=8.2.0

# Database (for app module compatibility)
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Web framework (for API and health checks in app module)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Templating (for app.ui)
jinja2>=3.1.2

# Authentication and security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
pynacl>=1.5.0

# Task queue
celery[redis]>=5.3.0
redis>=5.0.0

# WebSocket support
websockets>=12.0
